requirements.txt
Flask
rasterio
authlib
python-dotenv
